api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["kvfi <mail@ouafi.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
djangorestframework = "^3.15.1"
markdown = "^3.6"
django-filter = "^24.2"
django-oauth-toolkit = "^2.3.0"
psycopg2 = "^2.9.9"
django-extensions = "^3.2.3"
python-dotenv = "^1.0.1"
django-cors-headers = "^4.3.1"
flake8 = "^7.0.0"
gunicorn = "^21.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
